вынес connect из компонентов в отдельные файлы
