+++ src/components/avatar/avatar.connect.js
@@ -1,0 +1,9 @@
+import {connect} from 'react-redux';
+
+
+const mapStateToProps = ({user}) => ({
+  isAuthenticated: user.isAuthenticated,
+});
+
+
+export default connect(mapStateToProps);
