+++ src/components/genre-list/genre-list.connect.js
@@ -1,0 +1,16 @@
+import {connect} from 'react-redux';
+import {changeGenre} from 'store/actions/movies';
+import {getGenres} from 'store/selector';
+
+
+const mapStateToProps = ({movies, load}) => ({
+  currentGenre: movies.currentGenre,
+  genres: getGenres(load.movies),
+});
+
+const mapDispatchToProps = {
+  changeGenreAction: changeGenre,
+};
+
+
+export default connect(mapStateToProps, mapDispatchToProps);
