+++ src/components/movie/movie.connect.js
@@ -1,0 +1,9 @@
+import {connect} from 'react-redux';
+
+
+const mapStateToProps = ({load}) => ({
+  movies: load.movies,
+});
+
+
+export default connect(mapStateToProps);
