+++ src/components/my-list/my-list.connect.js
@@ -1,0 +1,10 @@
+import {connect} from 'react-redux';
+import {getMyList} from 'store/selector';
+
+
+const mapStateToProps = ({load}) => ({
+  movies: getMyList(load.movies),
+});
+
+
+export default connect(mapStateToProps);
