+++ src/components/promo/promo.connect.js
@@ -1,0 +1,9 @@
+import {connect} from 'react-redux';
+
+
+const mapStateToProps = ({load}) => ({
+  moviePromo: load.moviePromo,
+});
+
+
+export default connect(mapStateToProps);
