+++ src/components/sign-in/sign-in.connect.js
@@ -1,0 +1,10 @@
+import {connect} from 'react-redux';
+import {login} from 'store/actions/async';
+
+
+const mapDispatchToProps = {
+  loginAction: login,
+};
+
+
+export default connect(null, mapDispatchToProps);
